+++ History.md
@@ -2,6 +2,7 @@ Unreleased
 ==================
 
  * added; entity use code on accounts
+ * added; invoice previews
 
 1.1.5 (stable) / 2014-07-30
 ==================